fix polynorm example; remove DP dependency and lagrange polys from MU
