+++ Project.toml
@@ -5,7 +5,6 @@ version = "0.1.0"
 
 [deps]
 Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
-DynamicPolynomials = "7c1d4256-1411-5781-91ec-d7bc3513ac07"
 GenericLinearAlgebra = "14197337-ba66-59df-a3e3-ca00e7dcff7a"
 IterativeSolvers = "42fd0dbc-a981-5370-80f2-aaf504508153"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
@@ -20,7 +19,6 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
 Combinatorics = "1"
-DynamicPolynomials = "0.3"
 GenericLinearAlgebra = "0.2"
 IterativeSolvers = "0.8"
 LinearMaps = "2.7"
@@ -35,6 +33,7 @@ DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+DynamicPolynomials = "7c1d4256-1411-5781-91ec-d7bc3513ac07"
 GSL = "92c85e6c-cbff-5e0c-80f7-495c94daaecd"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 PolyJuMP = "ddf597a6-d67e-5340-b84c-e37d84115374"
@@ -43,4 +42,4 @@ SemialgebraicSets = "8e049039-38e8-557d-ae3a-bc521ccf6204"
 SumOfSquares = "4b9e565b-77fc-50a5-a571-1244f986bda1"
 
 [targets]
-test = ["CSV", "DataFrames", "DataStructures", "DelimitedFiles", "Distributions", "GSL", "JuMP", "PolyJuMP", "Random", "SemialgebraicSets", "SumOfSquares"]
+test = ["CSV", "DataFrames", "DataStructures", "DelimitedFiles", "Distributions", "DynamicPolynomials", "GSL", "JuMP", "PolyJuMP", "Random", "SemialgebraicSets", "SumOfSquares"]
